Introduced fbr_id_t
It is a 128bit integer composed of struct fbr_fiber * and uint64_t
counter. Now it is used as fiber identifier instead of a direct pointer.

This was done to solve ABA problem when fiber A gets reclaimed and
fiber B is created some other fiber holding a pointer to A may think
that A is still alive resulting in confusion.

Now fbr_id_t uniquely identifies a fiber despite of it being reclaimed
and created again at same address.