More rethinking and reworking of the event facility
fbr_ev_wait had a bug that lead to event loss when more than one event
arrived if a fiber was not scheduled to run soon enough. I fixed it, now
fbr_ev_wait catches all events that may arrive while waiting, each event
will have a flag indicating whether it has arrived or not.

Also reworked the allocation of the so-called ``items'', which are
elements of delayed execution queues (e.g. mutex pending queue, or
scheduled fibers queue). Now they are allocated on stack, include the
head of tailq they are currently in (if any) and are removed from the
queues only by a destructor. pending_async_callback does not remove
anything from the queue from now on. A unit test was added to match this
scenario.

To satisfy the need of a stack-based destructors a new destructor api
has been introduced. Destructor structure is exposed in public api, so
it can be allocated on stack as well as on heap (or in more obscure
places).